% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{GetFolderPath}
\alias{GetFolderPath}
\title{Get the OneDrive Path for WSFL Data}
\usage{
GetFolderPath(CALENDARYEAR, TERM)
}
\arguments{
\item{CALENDARYEAR}{An integer specifying the calendar year (e.g., 2023, 2024).}

\item{TERM}{An integer specifying the school term (e.g., 1, 2, 3, or 4).}
}
\value{
A string containing the full file path to the relevant WSFL data directory.
}
\description{
Constructs the file path for WSFL school class lists based on the selected calendar year and term.
The function dynamically detects the OneDrive path and builds the directory structure accordingly.
}
\examples{
\dontrun{
folder_path <- GetFolderPath(2024, 3)
print(folder_path)
}

}
