% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{GetWSFLAzureConnection}
\alias{GetWSFLAzureConnection}
\title{Connect to the WSFL Azure SQL Database}
\usage{
GetWSFLAzureConnection(
  Username = Sys.getenv("WSNZDBUSER"),
  Password = Sys.getenv("WSNZDBPASS")
)
}
\arguments{
\item{Username}{A string specifying the database username. Defaults to `Sys.getenv("WSNZDBUSER")`.}

\item{Password}{A string specifying the database password. Defaults to `Sys.getenv("WSNZDBPASS")`.}
}
\value{
A DBI connection object (`DBI::dbConnect`) if successful.
}
\description{
Establishes an ODBC connection to the WSFL Azure SQL Server database.
}
\examples{
\dontrun{
con <- GetWSFLAzureConnection()
DBI::dbListTables(con)
DBI::dbDisconnect(con)
}

}
